ErrorLogger-WebApp-ASP.NET Core
The application is aimed at software engineers and it will allow them to log error messages in one central place. The goal of the application is to not only log messages, but to also provide analysis of the collected data.

The system we are building consists of 3 parts:
1.	A DLL(library)
2.	A Rest Service
3.	A Website

System Use
When a developer want to use the error logger, they would reference the library (DLL) in their (host) application.
The library would be configured at the application start and it would be used to handle errors in the host application.

When the host application needs to log an error, it would make a call into the library.
The library would then send the error message into the REST service.
The service would be responsible for logging the error into the database.

Finally, when the developer is interested in reviewing logged error messages, they would log into the website and access all of the error logs that were saved in the DB.